requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
pandas==1.3.3
appwrite==0.8.0
schedule==1.1.0
python-dotenv==0.19.0
pyTelegramBotAPI==4.1.1
openpyxl==3.0.9